@@ -20,21 +20,21 @@
       "exclude": ["/tests", "./github"]
     },
     "require": {
-        "php":                                 "^5.6|^7.0",
-        "symfony/framework-bundle":            "~2.8|~3.0|~4.0",
-        "symfony/twig-bundle":                 "~2.8|~3.0|~4.0",
-        "symfony/form":                        "~2.8|~3.0|~4.0",
-        "symfony/asset":                       "~2.8|~3.0|~4.0",
-        "components/elfinder":                 "~3",
-        "studio-42/elfinder":                  "~2.1",
-        "robloach/component-installer":        "~0.2"
+        "php":                                 "^7.2",
+        "ext-json":                            "*",
+        "symfony/framework-bundle":            "^4.4 || ^5.0",
+        "symfony/twig-bundle":                 "^4.4 || ^5.0",
+        "symfony/form":                        "^4.4 || ^5.0",
+        "symfony/asset":                       "^4.4 || ^5.0",
+        "studio-42/elfinder":                  "~2.1"
     },
     "require-dev" : {
-        "symfony/phpunit-bridge": "^4.0",
-        "matthiasnoback/symfony-config-test":  "^3.0",
-        "matthiasnoback/symfony-dependency-injection-test": "^2",
-        "symfony/finder": "~3.0|~4.0",
-        "satooshi/php-coveralls": "~2"
+        "symfony/phpunit-bridge": "^4.1 || ^5.0",
+        "matthiasnoback/symfony-config-test":  "^4.0",
+        "matthiasnoback/symfony-dependency-injection-test": "^4.1",
+        "symfony/finder": "^4.4 || ^5.0",
+        "php-coveralls/php-coveralls": "^2.0",
+        "phpunit/phpunit": "^8.5"
     },
     "suggest": {
         "helios-ag/fm-tinymce-bundle": "FMTinyMCEBundle WYSIWYG Editor Bundle",
@@ -51,7 +51,7 @@
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "9.0-dev"
+            "dev-master": "10.0-dev"
         }
     }
 }
