@@ -17,11 +17,8 @@ class Configuration implements ConfigurationInterface
     public function getConfigTreeBuilder()
     {
         $treeBuilder = new TreeBuilder('fm_elfinder');
-        if (\method_exists($treeBuilder, 'getRootNode')) {
-            $rootNode = $treeBuilder->getRootNode();
-        } else {
-            $rootNode = $treeBuilder->root('fm_elfinder');
-        }
+        $rootNode    = $treeBuilder->getRootNode();
+
         $rootNode
             ->fixXmlConfig('instance')
             ->children()
@@ -40,7 +37,6 @@ class Configuration implements ConfigurationInterface
                             ->scalarNode('editor_template')->defaultNull()->end()
                             ->booleanNode('fullscreen')->defaultTrue()->end()
                             ->scalarNode('theme')->defaultValue('smoothness')->end() // jQuery UI theme name
-                            ->booleanNode('include_assets')->defaultTrue()->end()
                             ->scalarNode('tinymce_popup_path')->defaultValue('')->end()
                             ->booleanNode('relative_path')->defaultTrue()->end()
                             ->scalarNode('path_prefix')->defaultValue('/')->end()
@@ -431,7 +427,7 @@ class Configuration implements ConfigurationInterface
     }
 
     /**
-     * @return \Symfony\Component\Config\Definition\Builder\NodeDefinition the plugins node
+     * @return NodeDefinition the plugins node
      */
     private function createPluginsNode()
     {
@@ -444,7 +440,7 @@ class Configuration implements ConfigurationInterface
     }
 
     /**
-     * @return \Symfony\Component\Config\Definition\Builder\NodeDefinition the bind node
+     * @return NodeDefinition the bind node
      */
     private function createBindsNode()
     {
@@ -457,7 +453,7 @@ class Configuration implements ConfigurationInterface
     }
 
     /**
-     * @return \Symfony\Component\Config\Definition\Builder\NodeDefinition the bind node
+     * @return NodeDefinition the bind node
      */
     private function createDriverOptionsNode()
     {
@@ -470,11 +466,9 @@ class Configuration implements ConfigurationInterface
     }
 
     /**
-     * Creates a node.
-     *
      * @param string $name the node name
      *
-     * @return \Symfony\Component\Config\Definition\Builder\NodeDefinition the node
+     * @return NodeDefinition the node
      */
     private function createNode($name)
     {
