@@ -5,18 +5,20 @@ namespace FM\ElfinderBundle\Tests\DependencyInjection;
 use FM\ElfinderBundle\DependencyInjection\FMElfinderExtension;
 use FM\ElfinderBundle\DependencyInjection\Configuration;
 use Matthias\SymfonyDependencyInjectionTest\PhpUnit\AbstractExtensionConfigurationTestCase;
+use Symfony\Component\Config\Definition\ConfigurationInterface;
+use Symfony\Component\DependencyInjection\Extension\ExtensionInterface;
 
 /**
  * Class ConfigurationLoadTest.
  */
 class ConfigurationLoadTest extends AbstractExtensionConfigurationTestCase
 {
-    protected function getContainerExtension()
+    protected function getContainerExtension(): ExtensionInterface
     {
         return new FMElfinderExtension();
     }
 
-    protected function getConfiguration()
+    protected function getConfiguration(): ConfigurationInterface
     {
         return new Configuration();
     }
@@ -38,7 +40,6 @@ class ConfigurationLoadTest extends AbstractExtensionConfigurationTestCase
                     'theme'              => 'smoothness',
                     'editor_template'    => 'Elfinder/editor.html.twig',
                     'fullscreen'         => false,
-                    'include_assets'     => false,
                     'cors_support'       => false,
                     'tinymce_popup_path' => '/pop-up',
                     'relative_path'      => false,
