@@ -2,23 +2,18 @@
 
 namespace FM\ElfinderBundle\Tests\Loader;
 
+use FM\ElfinderBundle\Configuration\ElFinderConfigurationProviderInterface;
 use FM\ElfinderBundle\Loader\ElFinderLoader;
 
 class ElFinderLoaderTest extends \PHPUnit\Framework\TestCase
 {
-    /**
-     * @var ElFinderLoader
-     */
     protected $loader;
 
-    /**
-     * @var
-     */
     protected $configuratorMock;
 
-    public function setUp()
+    public function setUp(): void
     {
-        $this->configuratorMock = $this->createMock('FM\ElfinderBundle\Model\ElFinderConfigurationProviderInterface');
+        $this->configuratorMock = $this->createMock(ElFinderConfigurationProviderInterface::class);
         $this->configuratorMock->expects($this->any())
                                ->method('getConfiguration')
                                ->will($this->returnValue(array('parameters' => array())));
